requirements.txt
autopep8==1.4.4
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.7.1
Django==3.1.5
django-allauth==0.44.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-summernote==0.8.11.6
idna==2.10
image==1.5.33
oauthlib==3.1.0
pep8==1.7.1
Pillow==8.3.1
pycodestyle==2.5.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
stripe==2.60.0
urllib3==1.26.4